+++ changelog.md
@@ -3,6 +3,7 @@
 ## Upcoming
 * Added `hardcore` condition
   * Condition is fulfilled if the world is in hardcore mode
+* Enhanced errors when deserializing registry objects
 
 ## 5.0.13
 * Fixed `damaged` event not getting triggered with no attackers (e.g. fall damage)